Textrenderer initial refactoring
